+++ kroxylicious-filters/kroxylicious-record-validation/src/main/java/io/kroxylicious/filter/validation/validators/bytebuf/AvroSchemaBytebufValidator.java
@@ -23,7 +23,6 @@ import io.apicurio.registry.resolver.DefaultSchemaResolver;
 import io.apicurio.registry.resolver.SchemaResolver;
 import io.apicurio.registry.resolver.strategy.ArtifactReference;
 import io.apicurio.schema.validation.avro.AvroSchemaParser;
-import io.apicurio.schema.validation.avro.AvroValidationResult;
 import io.apicurio.schema.validation.avro.AvroValidator;
 
 import io.kroxylicious.filter.validation.config.SchemaValidationConfig.WireFormatVersion;
@@ -52,9 +51,7 @@ class AvroSchemaBytebufValidator extends AbstractSchemaBytebufValidator {
     protected CompletionStage<Result> doValidate(ByteBuffer buffer) {
         try {
             GenericRecord record = deserialize(buffer);
-            AvroValidationResult avroValidationResult = avroValidator.validateByArtifactReference(record);
-            return avroValidationResult.success() ? Result.VALID_RESULT_STAGE
-                    : CompletableFuture.completedFuture(new Result(false, avroValidationResult.toString()));
+            return toResult(avroValidator.validateByArtifactReference(record));
         }
         catch (IOException e) {
             return CompletableFuture.completedFuture(new Result(false, "Failed to deserialize Avro record: " + e.getMessage()));