+++ kroxylicious-filters/kroxylicious-record-validation/src/main/java/io/kroxylicious/filter/validation/validators/bytebuf/ProtobufSchemaBytebufValidator.java
@@ -24,7 +24,6 @@ import io.apicurio.registry.resolver.SchemaResolver;
 import io.apicurio.registry.resolver.strategy.ArtifactReference;
 import io.apicurio.registry.utils.protobuf.schema.ProtobufSchema;
 import io.apicurio.schema.validation.protobuf.ProtobufSchemaParser;
-import io.apicurio.schema.validation.protobuf.ProtobufValidationResult;
 import io.apicurio.schema.validation.protobuf.ProtobufValidator;
 
 import io.kroxylicious.filter.validation.config.SchemaValidationConfig.WireFormatVersion;
@@ -93,9 +92,7 @@ class ProtobufSchemaBytebufValidator extends AbstractSchemaBytebufValidator {
             byte[] bytes = new byte[buffer.remaining()];
             buffer.get(bytes);
             DynamicMessage message = DynamicMessage.parseFrom(messageDescriptor, bytes);
-            ProtobufValidationResult protobufValidationResult = protobufValidator.validateByArtifactReference(message);
-            return protobufValidationResult.success() ? Result.VALID_RESULT_STAGE
-                    : CompletableFuture.completedFuture(new Result(false, protobufValidationResult.toString()));
+            return toResult(protobufValidator.validateByArtifactReference(message));
         }
         catch (InvalidProtocolBufferException e) {
             return CompletableFuture.completedFuture(new Result(false, "Failed to parse Protobuf message: " + e.getMessage()));